LLM/requirements.txt
flask==2.1.2
pandas==2.1.2
psycopg2==2.9.9
openai==0.27.5
python-dotenv==1.0.0
sqlalchemy==2.0.23
Werkzeug==2.2.2